+++ tests/testthat/test-ichimoku.R
@@ -120,3 +120,7 @@ test_that("is.ichimoku ok", {
   expect_true(is.ichimoku(cloud))
   expect_false(is.ichimoku(sample_ohlc_data))
 })
+
+test_that(".ichimoku ok", {
+  expect_identical(.ichimoku(sample_ohlc_data, ticker = "TKR"), cloud)
+})