+++ NEWS.md
@@ -1,9 +1,16 @@
-# ichimoku 1.2.5.13
+# ichimoku 1.2.99
+
+#### New features:
+
+* This is a development preview of version 1.3.
+* Subplots now feature on the same chart and values can be read via the infotip in `iplot()`.
+* R/S-type indicators now an option for `oanda_chart()` and `oanda_studio()`.
+* `.ichimoku()` introduced as a faster technical utility version of `ichimoku()` for use when data is already in the required format. 
 
 #### Updates:
 
-* Dependency on C++ compiler (and 'cpp11' package) retired.
-* Internal rolling min/max/mean functions re-implemented in C for performance.
+* Dependency on C++ compiler, 'cpp11'  and 'gtable' packages retired.
+* Internal rolling min/max/mean functions re-implemented directly in C.
 * `ichimoku()`, `look()`, `as.data.frame()` and `as_tibble()` now amongst functions which call C code internally.
 * Fixes `oanda_stream()` for when multiple bid/ask liquidity levels are returned in the stream.
 * Fixes `archive()` to work correctly on R-oldrel (4.0).